Added confirmation to trash component Confirm before deleting/restoring ref #65

+++ angular-client/src/app/components/dashboard/trash/trash.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { TrashService } from '../../../services/trash.service';
 import { AuthenticationService } from '../../../services/authentication.service';
+import { MatDialog } from '@angular/material';
+import { Observable } from 'rxjs/Observable';
+import { ConfirmModalComponent } from '../../modals/confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'app-trash',
@@ -9,37 +12,50 @@ import { AuthenticationService } from '../../../services/authentication.service'
 })
 export class TrashComponent implements OnInit {
 
-  // TODO replace with db data
   items = [];
 
-  constructor(public authService: AuthenticationService, private trashService: TrashService) { }
+  constructor(public authService: AuthenticationService,
+    private trashService: TrashService,
+    public dialog: MatDialog) { }
 
   ngOnInit() {
     this.loadTrashRecords();
   }
 
   delete(itemModel: String, itemID: String) {
-    this.trashService.deletePermanently(itemModel, itemID)
-      .subscribe(data => {
-        console.log(data);
-        this.loadTrashRecords();
-      });
+    this.confirmModal("Are you sure you want to permanently delete this record?").subscribe(result => {
+      if (result) {
+        this.trashService.deletePermanently(itemModel, itemID)
+          .subscribe(data => {
+            console.log(data);
+            this.loadTrashRecords();
+          });
+      }
+    });
   }
 
   restore(itemModel: String, itemID: String) {
-    this.trashService.restore(itemModel, itemID)
-      .subscribe(data => {
-        console.log(data);
-        this.loadTrashRecords();
-      });
+    this.confirmModal("Are you sure you want to restore this record?").subscribe(result => {
+      if (result) {
+        this.trashService.restore(itemModel, itemID)
+          .subscribe(data => {
+            console.log(data);
+            this.loadTrashRecords();
+          });
+      }
+    });
   }
 
   deleteAll() {
-    this.trashService.deleteAll()
-      .subscribe((data: [any]) => {
-        console.log(data);
-        this.loadTrashRecords();
-      });
+    this.confirmModal("Are you sure you want to permanently delete all records in the trash bin?").subscribe(result => {
+      if (result) {
+        this.trashService.deleteAll()
+          .subscribe((data: [any]) => {
+            console.log(data);
+            this.loadTrashRecords();
+          });
+      }
+    });
   }
 
   loadTrashRecords(): void {
@@ -47,6 +63,15 @@ export class TrashComponent implements OnInit {
       .subscribe((data: [any]) => {
         this.items = data;
       });
+  }
+
+  confirmModal(message): Observable<any> {
+    const dialogRef = this.dialog.open(ConfirmModalComponent, {
+      width: '250px',
+      data: { message: message },
+      disableClose: false
+    });
+    return dialogRef.afterClosed();
   }
 
 }
